// webapp/src/shared/generated/models/CreateChatroomDto.ts
/* tslint:disable */
/* eslint-disable */
/**
 * transcendence-app
 * The transcendence-app API description
 *
 * The version of the OpenAPI document: 1.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { exists, mapValues } from '../runtime';
/**
 * 
 * @export
 * @interface CreateChatroomDto
 */
export interface CreateChatroomDto {
    /**
     * 
     * @type {string}
     * @memberof CreateChatroomDto
     */
    name: string;
    /**
     * 
     * @type {string}
     * @memberof CreateChatroomDto
     */
    password: string | null;
    /**
     * 
     * @type {string}
     * @memberof CreateChatroomDto
     */
    confirmationPassword: string | null;
}

/**
 * Check if a given object implements the CreateChatroomDto interface.
 */
export function instanceOfCreateChatroomDto(value: object): boolean {
    let isInstance = true;
    isInstance = isInstance && "name" in value;
    isInstance = isInstance && "password" in value;
    isInstance = isInstance && "confirmationPassword" in value;

    return isInstance;
}

export function CreateChatroomDtoFromJSON(json: any): CreateChatroomDto {
    return CreateChatroomDtoFromJSONTyped(json, false);
}

export function CreateChatroomDtoFromJSONTyped(json: any, ignoreDiscriminator: boolean): CreateChatroomDto {
    if ((json === undefined) || (json === null)) {
        return json;
    }
    return {
        
        'name': json['name'],
        'password': json['password'],
        'confirmationPassword': json['confirmationPassword'],
    };
}

export function CreateChatroomDtoToJSON(value?: CreateChatroomDto | null): any {
    if (value === undefined) {
        return undefined;
    }
    if (value === null) {
        return null;
    }
    return {
        
        'name': value.name,
        'password': value.password,
        'confirmationPassword': value.confirmationPassword,
    };
}